Refactor to use new top-level `PackageIcon` component (#60628)
- removes PackageIcon from EPM section
- refactors code to use new top-level `PackageIcon` component
