+++ x-pack/plugins/ingest_manager/public/applications/ingest_manager/sections/epm/components/index.ts
@@ -3,5 +3,3 @@
  * or more contributor license agreements. Licensed under the Elastic License;
  * you may not use this file except in compliance with the Elastic License.
  */
-
-export { PackageIcon } from './package_icon';
